requirements.txt
aniso8601==8.0.0
astroid==2.3.2
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
pkg-resources==0.0.0
pylint==2.4.3
pymongo==3.9.0
pyrsistent==0.15.5
pytz==2019.3
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
python-dotenv
requests
authlib
